Reflect review comments as the code changes
