+++ iml-gui/crate/src/lib.rs
@@ -658,64 +658,3 @@ pub fn run() {
 
     log!("App started.");
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::test_utils::fixtures;
-    use iml_wire_types::{
-        db::OstPoolOstsRecord,
-        warp_drive::{ArcCache, ArcValuesExt, Cache},
-    };
-    use std::sync::Arc;
-
-    #[test]
-    fn test_cache_conversions() {
-        let c0: Cache = fixtures::get_cache();
-        let c1: ArcCache = (&c0).into(); // From<&Cache> for ArcCache
-        let c0_again: Cache = (&c1).into(); // From<&ArcCache> for Cache
-
-        let mut c2: ArcCache = c1.clone();
-        let mut c3: ArcCache = c2.clone();
-
-        let rec1 = Arc::new(OstPoolOstsRecord {
-            id: 1,
-            ostpool_id: 1,
-            managedost_id: 1,
-        });
-        let rec2 = Arc::new(OstPoolOstsRecord {
-            id: 2,
-            ostpool_id: 2,
-            managedost_id: 2,
-        });
-        let rec18 = Arc::clone(&c1.ost_pool_osts.get(&18).unwrap());
-        let rec19 = Arc::clone(&c1.ost_pool_osts.get(&19).unwrap());
-
-        c2.ost_pool_osts.insert(1, Arc::clone(&rec1));
-        c3.ost_pool_osts.insert(2, Arc::clone(&rec2));
-
-        // The entries to c2 and c3 are added independently despite sharing the same "body"
-        assert_eq!(
-            c1.ost_pool_osts,
-            im::hashmap!(18 => Arc::clone(&rec18), 19 => Arc::clone(&rec19))
-        );
-        assert_eq!(
-            c2.ost_pool_osts,
-            im::hashmap!(18 => Arc::clone(&rec18), 19 => Arc::clone(&rec19), 1 => rec1)
-        );
-        assert_eq!(
-            c3.ost_pool_osts,
-            im::hashmap!(18 => Arc::clone(&rec18), 19 => Arc::clone(&rec19), 2 => rec2)
-        );
-        // the original cache and the cache - conversion result - should be equal
-        assert_eq!(c0, c0_again);
-    }
-
-    #[test]
-    fn test_arc_values() {
-        let cache: ArcCache = (&fixtures::get_cache()).into();
-
-        let osts: Vec<&OstPoolOstsRecord> = cache.ost_pool_osts.values().map(|x| &**x).collect();
-        let osts2: Vec<&OstPoolOstsRecord> = cache.ost_pool_osts.arc_values().collect();
-        assert_eq!(osts, osts2);
-    }
-}
