+++ iml-wire-types/Cargo.toml
@@ -7,7 +7,7 @@ authors = ["IML Team <iml@whamcloud.com>"]
 edition = "2018"
 
 [dependencies]
-serde = { version = "1", features = ["derive", "rc"] }
+serde = { version = "1", features = ["derive"] }
 serde_json = "1.0"
 tokio-postgres = { version = "0.5.1", optional = true }
 postgres-types = { version = "0.1.0", optional = true }
